refactor(web): presentation-agnostic speed-tiers content + lifted toggle
Split SpeedTiersPanel into three reusable, presentation-agnostic pieces
(SpeedTiersFieldControls, SpeedTiersTable, SpeedTiersModifiers) driven by an
external toggle/setToggle, plus the shared calc helpers, in a new
speed-tiers-content module. The panel becomes a thin side-pane container.

WHY: a forthcoming speed-tiers dialog needs to reuse the identical math + UI +
state as the side pane. Lifting the toggle into TeamWorkspaceV2 (one
useSpeedTiersToggle instance forwarded to both render sites) means switching
between presentations never resets weather/modifiers/sort. The table also gains
an optional nameFilter applied to the sorted rows before the group-separator
pass so base-speed headers attach to the correct visible rows.

getTeamFastestSpeed stays exported from speed-tiers-panel.tsx (imported by
team-workspace.tsx). Markup/classNames preserved verbatim.
